refactor: useComment 훅 분리
